moved ToolsAndEulaPage to the ui folder as it is a shared component and does not belong to NotebooksWizard
